fix(clean-tests): refactoring directories
